webapp/src/app/system-status/system-status.component.spec.ts
import {
  beforeEach,
  beforeEachProviders,
  describe,
  expect,
  it,
  inject,
} from '@angular/core/testing';
import { ComponentFixture, TestComponentBuilder } from '@angular/compiler/testing';
import { Component } from '@angular/core';
import { By } from '@angular/platform-browser';
import { SystemStatusComponent } from './system-status.component';

describe('Component: SystemStatus', () => {
  let builder: TestComponentBuilder;

  beforeEachProviders(() => [SystemStatusComponent]);
  beforeEach(inject([TestComponentBuilder], function (tcb: TestComponentBuilder) {
    builder = tcb;
  }));

  it('should inject the component', inject([SystemStatusComponent],
      (component: SystemStatusComponent) => {
    expect(component).toBeTruthy();
  }));

  it('should create the component', inject([], () => {
    return builder.createAsync(SystemStatusComponentTestController)
      .then((fixture: ComponentFixture<any>) => {
        let query = fixture.debugElement.query(By.directive(SystemStatusComponent));
        expect(query).toBeTruthy();
        expect(query.componentInstance).toBeTruthy();
      });
  }));
});

@Component({
  selector: 'test',
  template: `
    <app-system-status></app-system-status>
  `,
  directives: [SystemStatusComponent]
})
class SystemStatusComponentTestController {
}

